Methods from ConversionUtils were moved into appropriate response classes.
