+++ src/main/java/io/knowledgeassets/myskills/server/community/RecommendedCommunityResponse.java
@@ -10,7 +10,9 @@ import lombok.EqualsAndHashCode;
 
 import java.util.List;
 
-import static io.knowledgeassets.myskills.server.util.ConversionUtils.*;
+import static io.knowledgeassets.myskills.server.user.UserSimpleResponse.convertUserListToUserSimpleResponseList;
+import static io.knowledgeassets.myskills.server.skill.SkillResponse.convertSkillListToSkillResponseList;
+import static io.knowledgeassets.myskills.server.community.LinkResponse.convertLinkListToLinkResponseList;
 
 @Data
 @EqualsAndHashCode(callSuper = true)
